+++ lista4/HipercubeGraph.hpp
@@ -5,10 +5,10 @@
 
 class HipercubeGraph : public Graph {
 private:
-    uint8_t cubeDimension;
+    size_t cubeDimension;
 
 public:
-    HipercubeGraph(uint8_t dimension);
+    HipercubeGraph(size_t dimension);
 
 private:
     static uint8_t hammingWeight(size_t value);
